@@ -1,8 +1,17 @@
 {
   "name": "react-bootstrap-date-picker",
-  "keywords": ["react", "react-component", "react-bootstrap", "bootstrap", "date picker", "calendar", "date", "picker"],
+  "keywords": [
+    "react",
+    "react-component",
+    "react-bootstrap",
+    "bootstrap",
+    "date picker",
+    "calendar",
+    "date",
+    "picker"
+  ],
   "main": "lib/index.js",
-  "version": "1.1.0",
+  "version": "2.0.0",
   "description": "React-Bootstrap based date picker.",
   "directories": {
     "test": "test"
@@ -12,39 +21,38 @@
     "react-bootstrap": "^0.28.1"
   },
   "devDependencies": {
-    "react": ">=0.14.0",
-    "react-bootstrap": "^0.28.1",
-    "react-dom": "^0.14.0",
     "assert": "*",
-    "co": "*",
-    "regenerator": "*",
-    "regenerator-loader": "*",
-    "babel":"^6.1.18",
-    "babel-cli":"6.2.0",
-    "babel-preset-es2015": "*",
-    "babel-preset-stage-1": "*",
-    "babel-preset-react": "*",
+    "babel": "^6.1.18",
+    "babel-cli": "^6.7.5",
     "babel-loader": "^6.2.0",
-    "isparta-loader": "*",
-    "expose-loader": "*",
-    "transform-loader": "*",
+    "babel-preset-es2015": "*",
+    "babel-preset-react": "*",
+    "babel-preset-stage-1": "*",
+    "co": "*",
     "envify": "*",
+    "es6-promise": "*",
+    "expose-loader": "*",
+    "isparta-loader": "*",
     "karma": "*",
+    "karma-browserstack-launcher": "*",
     "karma-chrome-launcher": "*",
+    "karma-coverage": "*",
+    "karma-coveralls": "*",
     "karma-firefox-launcher": "*",
     "karma-mocha": "*",
     "karma-sourcemap-loader": "*",
     "karma-webpack": "*",
-    "karma-browserstack-launcher": "*",
-    "es6-promise": "*",
-    "node-uuid": "*",
-    "karma-coveralls": "*",
-    "karma-coverage": "*",
-    "mocha-babel": "*",
-    "webpack": "*",
     "mocha": "*",
-    "webpack-dev-server": "*",
-    "react-addons-linked-state-mixin": "*"
+    "mocha-babel": "*",
+    "node-uuid": "*",
+    "react": "^15.0.1",
+    "react-bootstrap": "^0.28.1",
+    "react-dom": "^15.0.1",
+    "regenerator": "*",
+    "regenerator-loader": "*",
+    "transform-loader": "*",
+    "webpack": "*",
+    "webpack-dev-server": "*"
   },
   "scripts": {
     "example": "./node_modules/webpack-dev-server/bin/webpack-dev-server.js --content-base example/ --config ./example/webpack.example.config.js",
